backend/app/requirements.txt
Flask
flasgger
openai-whisper
librosa
numpy
torch
soundfile
requests
python-dotenv